Adding README and Removed Excess Calculation

+++ ConcentrationCalculation.py
@@ -1,50 +1,73 @@
 from mpl_toolkits.mplot3d import Axes3D
 import matplotlib.pyplot as plt
-from matplotlib import cm
 import pandas as pd
 import numpy as np
 import os
 import logging
 
+""" ConcentrationCalculation.py 
+Calculates the Needed values from the CSV.
+"""
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 class ConcentrationCalculation:
     def __init__(self) -> None:
-        self._fn = os.environ.get('CON_CSV')
-        self._df_con = pd.read_csv(self._fn)
-        self._df_x = self._df_con['x']
-        self._df_y = self._df_con['y']
-        self._df_z = self._df_con['z']
-        self._col_con = self._df_con['concentration']
+        self.__conMean = None
+        self.__conStd = None
+        self.__conSum = None
+
+        # Reads in Environment Variable for CSV path
+        self.__fn = os.environ.get('CON_CSV') 
+        self.__df_con = pd.read_csv(self.__fn)
+
+        # Segments data based on column names
+        self.__df_x = self.__df_con['x']
+        self.__df_y = self.__df_con['y']
+        self.__df_z = self.__df_con['z']
+        self.__col_con = self.__df_con['concentration']
+
         logger.info("Entered Constructor")
     
+    # For Mean of Concentration
     def calcMean(self):
-        _conMean = np.mean(self._col_con)
-        logger.info(_conMean)
-        return _conMean
+        if self.__conMean is None:
+            self.__conMean = np.mean(self.__col_con)
+            logger.info(self.__conMean)
+        return self.__conMean
 
+    # For Standard Deviation of Concentration
     def calcStdDev(self):
-        _conStd = np.std(self._col_con)
-        logger.info(_conStd)
-        return _conStd
+        if self.__conStd is None:
+            self.__conStd = np.std(self.__col_con)
+            logger.info(self.__conStd)
+        return self.__conStd
 
+    # For Sum of Concentration
     def calcSum(self):
-        _conSum = np.sum(self._col_con)
-        logger.info(_conSum)
-        return _conSum
+        if self.__conSum is None:
+            self.__conSum = np.sum(self.__col_con)
+            logger.info(self.__conSum)
+        return self.__conSum
 
+    # For Image of Points using a 3D Scatter Plot w/ Concentration for Color
     def calcImg(self):
-        fig = plt.figure(figsize = (10,10))
-        ax = plt.axes(111, projection='3d')
-        sctt = ax.scatter3D(self._df_x, self._df_y, self._df_z, c = self._col_con, s = 50)
-        plt.title("Aeris Scatter Plot")
-        ax.set_xlabel('X-axis', fontweight ='bold') 
-        ax.set_ylabel('Y-axis', fontweight ='bold') 
-        ax.set_zlabel('Z-axis', fontweight ='bold')
-        fig.colorbar(sctt, ax = ax, shrink = 0.5, aspect = 5)
-        logger.info(self._col_con.max())
-        plt.savefig('./static_images/Aeris_ScatterPlot.png')
+        if not (os.path.isfile('./static_images/Aeris_ScatterPlot.png')):
+            fig = plt.figure(figsize = (10,10))
+            ax = plt.axes(111, projection='3d')
+            sctt = ax.scatter3D(self.__df_x, self.__df_y, self.__df_z, c = self.__col_con, s = 50)
+            plt.title("Aeris Scatter Plot")
+            ax.set_xlabel('X-axis', fontweight ='bold') 
+            ax.set_ylabel('Y-axis', fontweight ='bold') 
+            ax.set_zlabel('Z-axis', fontweight ='bold')
+            cbar = fig.colorbar(sctt, ax = ax, shrink = 0.5, aspect = 5)
+            cbar.ax.set_ylabel('Concentration', rotation=270, labelpad=10)
+            logger.info(self.__col_con.max())
+
+            #Saves Images for Controller
+            plt.savefig('./static_images/Aeris_ScatterPlot.png')
+
+        #Returns Absolute Path for Controller
         return os.path.abspath('./static_images/Aeris_ScatterPlot.png')
